Asset polishing part 2 and movement
Made the cargo crates look better and added guns around the level for the player to pick up. I have only codded the guns to spin so far, so the inventory system will still need to be made.

Added code for the player to move using "WASD" and to let them look around with the mouse.

The level feels less empty now but I am not sure what else I should add as decoration as I'm worrying about overstimulating the player with random goofy objects.