Main Menu Scene
Added the scene "Main Menu Scene". In this scene, I have the player hiding behind a bush outside of the building and upon hitting a start button, the player will glide towards the door and the scene will change to the newly renamed "Main Level Scene". I also made the table smaller in the main scene.

TLDR: Added Menu with functioning "Start" and "Quit Game" buttons in addition to some shrubbery.  Made the table smaller in the level