Major File Restructuring/Refactoring
Made major changes to the file and script structure.

Almost everything that the player controls is now in PlayerController. The functions of handling the camera controlling, player movement, and weapon management are written up as their own functions that are referenced/run in Update().

Also did some work with the model so that Snake isn't just a weird white mannequin as well as added a sprint and a jump.