Hacking Minigame Update 2
Worked on means of switching to the hacking minigame. I got a basic switch working, but it is still rough around the edges, e.g., it is possible to switch to an already-hacked robot during the minigame.
Note: I realized that weapon switching and robot switching are both tied to the number keys which could cause some issues, so I will likely need to  rework one of the two.