Basics of the improved weapon system
New guns added, including my shotgun. Complete with the necessary changes to HoverAndRotate and PlayerMovement that make the system of adding new weapons to the possible list far easier.
New script that applies to all of the weapons that makes gives weapons a magazine size, current ammunition, a Boolean for whether or not the gun is full auto, as well as a randomized bullet spread function.

Some of these don't seem to be working properly as intended, however. Shotgun spread looks like it is somehow being influenced/limited by the direction of aim and all other weapons seem to now be tiny when picked up. Will try to fix this in next commit.