Weapons UI Added
I have implemented the weapon system into the Heads Up Display (HUD). Shows both the ammo count in the weapon and the total in reserve. If the ammo dips below 25%, the color of the weapon's ammo counter turns red.
I also modified the muzzleFlash particle systems as well as fixed the problem of the imported weapons being tiny when picked up.