Fixed Robot UI Ammo Counter/Added One + Win Screen Modification
Now you can view the ammo counter on the weapons of the robots. They still have infinite ammo, by the way. Just make sure you reload the magazines.
I also changed the scripts in general
Plus a few tweaks to the end screen to make it properly readable and visible.